airquality/requirements.txt
APScheduler==3.10.1
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
Django==4.2.2
djangorestframework==3.14.0
drf-spectacular==0.26.3
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
pycparser==2.21
PyMySQL==1.1.0
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
requests==2.31.0
rpds-py==0.8.8
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.3
